Remove base.py file from publishers
