+++ tests/wagtail_live/publishers/test_intervalpollingpublisher.py
@@ -7,8 +7,10 @@ from django.urls import resolve
 
 from tests.testapp.models import BlogPage
 from tests.utils import reload_urlconf
-from wagtail_live.publishers.base import PollingPublisherMixin
-from wagtail_live.publishers.polling import IntervalPollingPublisher
+from wagtail_live.publishers.polling import (
+    IntervalPollingPublisher,
+    PollingPublisherMixin,
+)
 from wagtail_live.utils import get_polling_interval
